+++ CHANGELOG.md
@@ -9,11 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Upload attachments (currently image-only)
 - Download attachmenst
+- Remove reply message before sending a message
+- Remove attachments before sending a message
+- Improved UI for replies in the chat
 
 ### Fixed
 
 - Reply message not being cleared when sending a message
 - Reply message not being cleared when switching channels
+- Replies not showing correctly
 
 ## [0.1.4] - 2022-07-16
 ### Added